Rename DashboardView/utils.ts to formatters (#1171)
Also, removes some low-value comments that only repeat file names.
