+++ assets/app/vue/views/DashboardView/formatters.ts
