+++ assets/app/vue/views/SignUpView/views/StepConfirmPlan/index.vue
@@ -7,7 +7,7 @@ import PlanCard from '@/components/PlanCard.vue';
 import { useI18n } from 'vue-i18n';
 import SignUpLayout from '../../components/SignUpLayout.vue';
 import { useSignUpFlowStore } from '../../stores/signUpFlowStore';
-import { formatLocalizedMonthlyPrice } from '@/views/DashboardView/utils';
+import { formatLocalizedMonthlyPrice } from '@/views/DashboardView/formatters';
 const DEFAULT_MONTHLY_PRICE = '$6';
 
 const { t } = useI18n();
